requirements.txt
PyYAML
colorama
